requirements.txt
fastapi
pandas
numpy
langchain
langchain-community
fastapi
uvicorn
python-dotenv
python-multipart
pillow
ipython
streamlit
